chore(@deip/auth-module)add current user to auth module - Move Current User to auth module - Remove roles from state - Refactor auth-module
